+++ src/Forms/Shared/Samples/Data/GenerateGeodatabaseReplica/GenerateGeodatabaseReplica.xaml
@@ -2,7 +2,7 @@
 <ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              xmlns:esriUI="clr-namespace:Esri.ArcGISRuntime.Xamarin.Forms;assembly=Esri.ArcGISRuntime.Xamarin.Forms"
-             x:Class="ArcGISRuntime.Samples.GenerateGeodatabase.GenerateGeodatabase">
+             x:Class="ArcGISRuntime.Samples.GenerateGeodatabaseReplica.GenerateGeodatabaseReplica">
     <Grid>
         <Grid.RowDefinitions>
             <RowDefinition Height="auto" />
