+++ src/WPF/ArcGISRuntime.WPF.Viewer/Samples/Data/GenerateGeodatabaseReplica/GenerateGeodatabaseReplica.jpg
