+++ src/WPF/ArcGISRuntime.WPF.Viewer/Samples/Data/GenerateGeodatabaseReplica/GenerateGeodatabaseReplica.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="ArcGISRuntime.WPF.Samples.GenerateGeodatabase.GenerateGeodatabase"
+﻿<UserControl x:Class="ArcGISRuntime.WPF.Samples.GenerateGeodatabaseReplica.GenerateGeodatabaseReplica"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
