+++ src/WinUI/ArcGISRuntime.WinUI.Viewer/Samples/Data/GenerateGeodatabaseReplica/GenerateGeodatabaseReplica.jpg
