+++ src/WinUI/ArcGISRuntime.WinUI.Viewer/Samples/Data/GenerateGeodatabaseReplica/GenerateGeodatabaseReplica.xaml
@@ -1,5 +1,5 @@
 ﻿<UserControl
-    x:Class="ArcGISRuntime.WinUI.Samples.GenerateGeodatabase.GenerateGeodatabase"
+    x:Class="ArcGISRuntime.WinUI.Samples.GenerateGeodatabaseReplica.GenerateGeodatabaseReplica"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:esriUI="using:Esri.ArcGISRuntime.UI.Controls">
